eng_learner_web_app/backend/requirements.txt
fastapi
pydantic
python-dotenv
google-generativeai
uvicorn
requests